Move AssociationWithPermissions public api model to AC Team
