+++ src/Api/AdminConsole/Public/Models/Request/GroupCreateUpdateRequestModel.cs
@@ -1,5 +1,4 @@
-﻿using Bit.Api.Auth.Models.Public.Request;
-using Bit.Core.AdminConsole.Entities;
+﻿using Bit.Core.AdminConsole.Entities;
 
 namespace Bit.Api.AdminConsole.Public.Models.Request;
